Set modern-installation config option to false
As of Poetry v1.4.1, packages were required to use modern installation (see: https://github.com/python-poetry/poetry/issues/7686). This is potentially problematic when relying on older versions of packages and for packages who are still using non-modern installers.

This commit adds an additional step to disable that modern-installation requirement, which should get the workflow working again.