change use of expand, and refactoring for inputs
This uses snakebids (BidsComponent) expand when we are directly accesssing bids
dataset inputs, and the regular snakemake expand when we are not.

Also refactors the input functions for the import rules, for cases where
we only deal with 1 input (seg, cropseg, and currently T1w), and for
cases where we process all the inputs (T2w). Eventually we should make
the T1w workflow also process all the inputs (as is done in the T2w
case), making rules generic to modality (do this in another PR).