Update snakebids to its latest version (v0.13.1) (#318)
* updated pyproject.toml to use the latest version of snakebids

* Using invalid_filters key in pybids_inputs breaks the pipeline, commenting out

* work with the latest version of bids_spec and avoid using relative paths

* generate a new bids function to remove session and subject dirs

* use snakebids create to work with the latest template and syntax

* add description for each dataset in the dry run test

* use a fallback definition for __file__ variable in run.py

* remove bids validation by default

* adds an enable-bids-validator flag instead of a skip flag (for backwards
compatibility)..

*  removes the core bids CLI args from config, as these are added
automatically with new snakebids

* change use of expand, and refactoring for inputs

This uses snakebids (BidsComponent) expand when we are directly accesssing bids
dataset inputs, and the regular snakemake expand when we are not.

Also refactors the input functions for the import rules, for cases where
we only deal with 1 input (seg, cropseg, and currently T1w), and for
cases where we process all the inputs (T2w). Eventually we should make
the T1w workflow also process all the inputs (as is done in the T2w
case), making rules generic to modality (do this in another PR).

---------

Co-authored-by: Ali Khan <alik@robarts.ca>