refactoring of the laplace beltrami rule (#371)
- splits into different rules for extracting the boundary, and getting
distances to src/sink masks
- those new rules now produce surface maps (for easier introspection),
which are passed along to the laplace beltrami rule
- the dentate AP src in bbhist hemi-L was problematic with existing
approach, since the PD src/sink were always closer than the AP src, so
no vertices were being labelled with AP src.
- in this version, we do things a little different:
  1. AP and PD are done independently now
  2. Instead of nearest neighbor, we use a distance threshold, which is
defined based on a signed distance transform of the mask, mapped to the
surface.
  3. Actually there is a minimum distance, and maximum distance, along
with the minimum number of vertices required in the src/sink. The
threshold is swept until the desired number is reached, to deal with
cases such as the above.

- TODO: the get_boundary_vertices script could be further improved to
perform connected components and picking the largest one, to avoid being
affected by holes or other defects..

- TODO: the laplace beltrami solver doesn't take as long to run now that
I am decimating alot, but if we want to further optimize things, I think
the step of altering the laplacian based on the boundary conditions, after it has been made a sparse matrix, is where the inefficiency lies (based on profiling). Could try to avoid setting those weights in the first place instead of setting it after the fact..

- NOTE: solve_laplace_beltrami_open_mesh() is kept as is (except for some
logging statements)